+++ C#/Day3/Logic/IBatteryChooser.cs
@@ -3,5 +3,5 @@
 public interface IBatteryChooser
 {
     int Choose2Batteries(string input);
-    double Choose12Batteries(string input);
+    long Choose12Batteries(string input);
 }
